Fetching now the data from Google, but connection gets interrupted, need Channels and a worker

+++ src/Amex/Response.hs
@@ -12,7 +12,7 @@ data Store = Store { name   :: Text
                    , zip    :: Int
                    , city   :: Text
                    , address:: Text
-                   , dist   :: Int
+                   , dist   :: Double
                    } deriving (Eq, Generic, Show)
 
 instance ToJSON Store where
